refactor: collapse turn integration tests into four domain entrypoints (#3876)
Move scenario bodies to test/support/integration/*_scenarios.dart and
keep four thin integration entrypoints (combat, diplomacy, economy,
spy/fog) so AC2 is satisfied without part fragments or 400-line cap
violations.
